deployment/requirements.cdk.txt
attrs==20.3.0
aws-cdk.assets==1.81.0
aws-cdk.aws-apigateway==1.81.0
aws-cdk.aws-applicationautoscaling==1.81.0
aws-cdk.aws-autoscaling-common==1.81.0
aws-cdk.aws-certificatemanager==1.81.0
aws-cdk.aws-cloudformation==1.81.0
aws-cdk.aws-cloudwatch==1.81.0
aws-cdk.aws-codeguruprofiler==1.81.0
aws-cdk.aws-dynamodb==1.81.0
aws-cdk.aws-docdb
aws-cdk.aws-ec2==1.81.0
aws-cdk.aws-ecr==1.81.0
aws-cdk.aws-ecr-assets==1.81.0
aws-cdk.aws-ecs
aws-cdk.aws-ecs-patterns
aws-cdk.aws-ecr==1.81.0
aws-cdk.aws-efs==1.81.0
aws-cdk.aws-elasticloadbalancingv2==1.81.0
aws-cdk.aws-events==1.81.0
aws-cdk.aws-iam==1.81.0
aws-cdk.aws-kms==1.81.0
aws-cdk.aws-lambda==1.81.0
aws-cdk.aws-logs==1.81.0
aws-cdk.aws-route53==1.81.0
aws-cdk.aws-s3==1.81.0
aws-cdk.aws-s3-assets==1.81.0
aws-cdk.aws-sns==1.81.0
aws-cdk.aws-sqs==1.81.0
aws-cdk.aws-ssm==1.81.0
aws-cdk.cloud-assembly-schema==1.81.0
aws-cdk.core==1.81.0
aws-cdk.custom-resources==1.81.0
aws-cdk.cx-api==1.81.0
aws-cdk.region-info==1.81.0
aws-cdk.aws-route53-targets==1.81.0
aws-cdk.aws-sns-subscriptions==1.81.0
python-dotenv==0.15.0
pydantic-kms-secrets==0.3.0
fastapi-pagination==0.5.1